+++ src/cells/SqrtCell.cpp
@@ -59,11 +59,13 @@ void SqrtCell::MakeBreakUpCells() {
 }
 
 void SqrtCell::Recalculate(AFontSize fontsize) const {
-  if (NeedsRecalculation(fontsize)) {
+  bool changed = false;
+  changed |= m_innerCell->RecalculateList(fontsize);
+  changed |= m_open->RecalculateList(fontsize);
+  changed |= m_close->RecalculateList(fontsize);
+
+  if (changed || NeedsRecalculation(fontsize)) {
     Cell::Recalculate(fontsize);
-    m_innerCell->RecalculateList(fontsize);
-    m_open->RecalculateList(fontsize);
-    m_close->RecalculateList(fontsize);
 
     m_width = m_innerCell->SumOfWidths() + Scale_Px(13) + 1;
     if (!IsBrokenIntoLines()) {